Removed unused web proxy and refactored name. Removed obsolete test folder in src
